TagBuilder: ignore duplicate tags
Change AddItemInternal to ignore items that are equal in type and
value. The check is done by strcmp() because TagPool might duplicate
items if their ref counter exceeds 255.

This is primarily a cosmetic fix for clients such as ncmpcpp, which do
in fact display tag duplicates to the user (e.g. "01 | 01" as a track
number). Duplicate tags might be present in audio files because
automatic taggers like Beets increase compatibility with various audio
players this way.

This patch preserves the previous behavior if multiple tags exist with
the same type but different value.